# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WhatsApp automation bot for Kantorku.id HRIS SaaS",
        "title": "Customer Service Agent Bot API",
        "contact": {
            "name": "Dealls Engineering",
            "url": "https://github.com/Sejutacita"
        },
        "version": "1.0"
    },
    "host": "api-dev.sejutacita.id",
    "basePath": "/v1/cs-agent-bot",
    "paths": {
        "/api/handoff/new-client": {
            "post": {
                "security": [
                    {
                        "X-Handoff-Secret": []
                    }
                ],
                "description": "Creates a new client record when a Business Development team member hands off a new prospect. Requires HMAC authentication using X-Handoff-Secret header.",
                "tags": [
                    "api"
                ],
                "summary": "Onboard New Client from BD Handoff",
                "parameters": [
                    {
                        "description": "New client handoff payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.HandoffCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Client onboarded successfully",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid HMAC signature",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
        "/cron/run": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually triggers the bot's cron job to process all clients and send scheduled WhatsApp messages. Requires GCP Cloud Scheduler OIDC token authentication.",
                "tags": [
                    "cron"
                ],
                "summary": "Trigger Cron Job",
                "responses": {
                    "200": {
                        "description": "Cron run completed successfully",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing OIDC token",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - cron run failed",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
        "/readiness": {
            "get": {
                "description": "Health check for service",
                "tags": [
                    "health"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
        "/webhook/checkin-form": {
            "post": {
                "description": "Processes check-in form submissions from clients. Marks the client as having replied to check-in and notifies the Account Executive via Telegram.",
                "tags": [
                    "webhook"
                ],
                "summary": "Process Check-in Form Submission",
                "parameters": [
                    {
                        "description": "Check-in form payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CheckinFormRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkin form processed successfully",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "Company not found",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.StandardResponse"
                        }
                    }
                }
            }
        },
        "/webhook/wa": {
            "post": {
                "description": "Receives incoming WhatsApp message events from HaloAI. Returns HTTP 200 immediately, then processes the message asynchronously in a background goroutine.",
                "tags": [
                    "webhook"
                ],
                "summary": "Receive WhatsApp Webhook",
                "parameters": [
                    {
                        "description": "WhatsApp webhook payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WAWebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook received (async processing)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apperror.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.CheckinFormRequest": {
            "type": "object",
            "required": [
                "company_id"
            ],
            "properties": {
                "company_id": {
                    "type": "string",
                    "example": "KTK-001"
                }
            }
        },
        "dto.HandoffCreateRequest": {
            "type": "object",
            "required": [
                "company_id",
                "company_name",
                "contract_end",
                "contract_months",
                "contract_start",
                "owner_name",
                "owner_telegram_id",
                "pic_name",
                "pic_wa",
                "segment"
            ],
            "properties": {
                "activation_date": {
                    "type": "string",
                    "example": "2024-01-15"
                },
                "company_id": {
                    "type": "string",
                    "example": "KTK-001"
                },
                "company_name": {
                    "type": "string",
                    "example": "PT Maju Bersama Sejahtera"
                },
                "contract_end": {
                    "type": "string",
                    "example": "2026-01-15"
                },
                "contract_months": {
                    "type": "integer",
                    "example": 12
                },
                "contract_start": {
                    "type": "string",
                    "example": "2024-01-15"
                },
                "owner_name": {
                    "type": "string",
                    "example": "Rina"
                },
                "owner_telegram_id": {
                    "type": "string",
                    "example": "-1001234567890"
                },
                "owner_wa": {
                    "type": "string",
                    "example": "6289876543210"
                },
                "pic_name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "pic_wa": {
                    "type": "string",
                    "example": "6281234567890"
                },
                "segment": {
                    "type": "string",
                    "example": "Mid"
                }
            }
        },
        "dto.WAWebhookRequest": {
            "type": "object",
            "properties": {
                "message_id": {
                    "type": "string",
                    "example": "msg_1234567890"
                },
                "message_type": {
                    "description": "text, image, voice, video, document",
                    "type": "string",
                    "example": "text"
                },
                "phone_number": {
                    "type": "string",
                    "example": "6281234567890"
                },
                "text": {
                    "type": "string",
                    "example": "Thank you for the reminder."
                },
                "timestamp": {
                    "type": "string",
                    "example": "2025-03-26T10:30:00Z"
                }
            }
        },
        "response.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "stackTrace": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apperror.FieldError"
                    }
                }
            }
        }
    }
}